refeat: GameService에서 UserService 분리
- GameService를 SRP에 따라 사용자 입력 기능을 UserService로 분리함(+관련 validation)
